+++ zuul/trigger/gerrit.py
@@ -138,22 +138,6 @@ class Gerrit(object):
         self.gerrit_connector.stop()
         self.gerrit_connector.join()
 
-    def report(self, change, message, action):
-        self.log.debug("Report change %s, action %s, message: %s" %
-                       (change, action, message))
-        if not change.number:
-            self.log.debug("Change has no number; not reporting")
-            return
-        if not action:
-            self.log.debug("No action specified; not reporting")
-            return
-        changeid = '%s,%s' % (change.number, change.patchset)
-        ref = 'refs/heads/' + change.branch
-        change._ref_sha = self.getRefSha(change.project.name,
-                                         ref)
-        return self.gerrit.review(change.project.name, changeid,
-                                  message, action)
-
     def _getInfoRefs(self, project):
         url = "%s/p/%s/info/refs?service=git-upload-pack" % (
             self.baseurl, project)